backend/app/api/api_v1/endpoints/chat.py
from typing import Any
from fastapi import APIRouter, Depends, HTTPException, WebSocket, WebSocketDisconnect, Query
from sqlalchemy.orm import Session
from jose import jwt, JWTError

from app import models, schemas
from app.api import deps
from app.core import config
from app.websocket.manager import manager
from app.services import chat as chat_service
from app.services.notification.service import notification_service

router = APIRouter()

@router.websocket("/ws/chat/{event_id}")
async def websocket_chat_endpoint(
    websocket: WebSocket,
    event_id: int,
    token: str = Query(...),
    db: Session = Depends(deps.get_db),
):
    # Authenticate via token
    user = None
    try:
        payload = jwt.decode(token, config.settings.SECRET_KEY, algorithms=[config.settings.ALGORITHM])
        user_id: str = payload.get("sub")
        if user_id is None:
            await websocket.close(code=4003) # Forbidden
            return
        user_id_int = int(user_id)
        # Verify user exists (simple check, or fetch full user)
        # user = db.query(models.User).filter(models.User.id == user_id_int).first()
    except (JWTError, ValueError):
        await websocket.close(code=4003)
        return

    # Check if user is participant (optional, but good practice)
    # participant = db.query(models.EventParticipant).filter(...)
    # if not participant: ...

    await manager.connect(websocket, event_id)
    
    try:
        while True:
            data = await websocket.receive_text()
            
            # Save message to DB
            # We strictly shouldn't use the 'db' session from Depends directly in async loop without care, 
            # but for this scale it acts as a quick prototype. 
            # Ideally use an async session or run in threadpool.
            # Here we just assume synchronous DB call is "okay" for MVP or wrap it.
            # For simplicity in this agent demo, we will persist synchronously.
            saved_msg = chat_service.save_message(db, user_id_int, event_id, data)
            
            message_data = {
                "id": saved_msg.id,
                "content": saved_msg.content,
                "sender_id": saved_msg.sender_id,
                "event_id": saved_msg.event_id,
                "created_at": str(saved_msg.created_at)
            }
            
            # Publish to Redis -> Broadcast to all
            await manager.publish_message(event_id, message_data)
            
            # Send Notification (In reality, we would fetch all participants and queue notifications)
            # For demo, let's just log/trigger for the sender purely as proof of concept, 
            # or ideally to OTHER participants.
            # Implementing "Notify others" requires fetching participants.
            # Skipping loop for now to focus on abstraction, but here is where it goes:
            # participants = db.query(models.EventParticipant)...
            # for p in participants:
            #    if p.user_id != user_id_int:
            #        await notification_service.send_new_message(p.user_id, "Sender Name", "Event Name", data)
            
    except WebSocketDisconnect:
        manager.disconnect(websocket, event_id)
    except Exception as e:
        # Handle other errors
        manager.disconnect(websocket, event_id)
